quadbot69/requirements.txt
discord==1.7.3
discord.py==1.7.3
googletrans==3.1.0a0
Pillow==8.0.1
requests==2.25.1
samp-client==3.0
english-words==1.0.3
bs4==0.0.1
beautifulsoup4==4.9.3
Pyrebase4==4.4.3
pytz==2021.1
